+++ projects/pastanaga-angular/src/lib/datetime/datetime.service.spec.ts
@@ -4,13 +4,15 @@ import { DATE_FORMAT, DateTimeService } from './datetime.service';
 
 import { TranslatePipe } from '../translate/translate.pipe';
 import { dates, mockTranslateEn, mockTranslateFr } from './test-data';
+import { PA_TRANSLATIONS } from '../translate/translate.model';
 
 describe('DateTimeService', () => {
     let service: DateTimeService;
 
     describe('en-US locale', () => {
         beforeEach(() => {
-            service = new DateTimeService(new TranslatePipe('en-US', { 'en-US': mockTranslateEn }), 'en-US');
+            PA_TRANSLATIONS.en_US = mockTranslateEn;
+            service = new DateTimeService(new TranslatePipe('en-US'), 'en-US');
         });
 
         it('should be defined', () => {
@@ -141,7 +143,8 @@ describe('DateTimeService', () => {
     describe('fr-FR locale', () => {
         beforeEach(() => {
             registerLocaleData(localeFr, 'fr-FR');
-            service = new DateTimeService(new TranslatePipe('fr-FR', { 'fr-FR': mockTranslateFr }), 'fr-FR');
+            PA_TRANSLATIONS['fr-FR'] = mockTranslateFr;
+            service = new DateTimeService(new TranslatePipe('fr-FR'), 'fr-FR');
         });
 
         describe('getFormattedDate method with human param', () => {