+++ projects/pastanaga-angular/src/lib/translate/translate.model.ts
@@ -1,3 +1,5 @@
+export const PA_TRANSLATIONS: Translation = {};
+
 export interface TranslationEntries {
     [id: string]: string | TranslationEntries;
 }