+++ projects/pastanaga-angular/src/lib/translate/translate.module.ts
@@ -1,7 +1,9 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { TranslatePipe } from './translate.pipe';
 import { TranslateDirective } from './translate.directive';
+import { PA_TRANSLATIONS, Translation } from './translate.model';
+import { mergeTranslations } from './translate.utils';
 
 @NgModule({
     imports: [CommonModule],
@@ -9,4 +11,12 @@ import { TranslateDirective } from './translate.directive';
     declarations: [TranslatePipe, TranslateDirective],
     providers: [TranslatePipe],
 })
-export class PaTranslateModule {}
+export class PaTranslateModule {
+    static addTranslations(translations: Translation[]): ModuleWithProviders<PaTranslateModule> {
+        mergeTranslations(translations);
+        return {
+            ngModule: PaTranslateModule,
+            providers: [],
+        };
+    }
+}